fix shakkala pip
shakkala official under pip and working
