--- MANIFEST.in
@@ -1,6 +1,0 @@
-include Shakkala.py
-include helper.py
-include __init__.py
-include dictionary/*
-include model/*
-include requirements/ *
