+++ requirements/publish_commands.txt
@@ -1,0 +1,3 @@
+python setup.py sdist
+python setup.py bdist_wheel
+twine upload dist/*
