+++ shakkala.egg-info/SOURCES.txt
@@ -1,26 +1,17 @@
-.DS_Store
 LICENSE.md
-MANIFEST.in
 README.md
-Shakkala.py
-__init__.py
-demo.py
-helper.py
 setup.py
-Shakkala.egg-info/PKG-INFO
-Shakkala.egg-info/SOURCES.txt
-Shakkala.egg-info/dependency_links.txt
-Shakkala.egg-info/requires.txt
-Shakkala.egg-info/top_level.txt
-dictionary/input_vocab_to_int.pickle
-dictionary/output_int_to_vocab.pickle
-model/middle_model.h5
-model/second_model6.h5
-model/simple_model.h5
-requirements/.DS_Store
-requirements/requirements.txt
+shakkala/Shakkala.py
+shakkala/__init__.py
+shakkala/demo.py
+shakkala/helper.py
 shakkala.egg-info/PKG-INFO
 shakkala.egg-info/SOURCES.txt
 shakkala.egg-info/dependency_links.txt
 shakkala.egg-info/requires.txt
 shakkala.egg-info/top_level.txt
+shakkala/dictionary/input_vocab_to_int.pickle
+shakkala/dictionary/output_int_to_vocab.pickle
+shakkala/model/middle_model.h5
+shakkala/model/second_model6.h5
+shakkala/model/simple_model.h5
