+++ shakkala.egg-info/top_level.txt
@@ -1,2 +1,1 @@
-Shakkala
-helper
+shakkala
